TypeError is Defined
From the base context of the program, this code represents only the handle exception error for the same priority value stack in the queue that can't be compared to another. When we consider their time of arrival, then the placed element between the equal order elements would be comparable to time.