Traversing the Graph Using BFA
BFA - Breath-First Approach to produce a path guaranteed to have the fewest nodes. There are two shortest paths between Abardeen and Perth when we disregard the road distances. As same as before, the actual result in this case will utmost depend on how the order the neighboring nodes are sorted.