Assigning an Infinite Loop on the Worker
With the def worker() function program, the role of a worker has been set as the producer and consumer, making new jobs to queue as it consumes data from the URL point that connects HTML hyperlinks with try except clauses to avoid unnecessary deadlocks.

One thing to add is that, when the depth moves from the accounted limit to exceeding hits, then it would handle a ClientError which usually arrives at a point where the job cannot fetch the HTML content.