Implemented Program Sequence Blocks
def callback() with its four parameters typically involves a variable named "message" that acts as a method decoder in a common computer language and  a print function indicating the received message.

With pika.Blockconnection(), we create a series of queue connection channels so that the QUEUE_NAME with string mailbox can callback on message and starts consuming it with a method shown at line 23.