Added Line Arguments for KafkaProducer Bootstrap
The .send() method implies asynchronous function. The program provides a future object that we await by invoking its blocking.get() function. On the consumer side, we'll be able to read the transmitted messages with the topic data science that encodes utf-8 by iterating it over the consumer using localhost:9092 bootstrap servers.